requirements.txt
absl-py==0.9.0
apex==0.1
astor==0.8.1
backcall==0.2.0
boto3==1.12.20
botocore==1.15.20
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.16.0
dataclasses==0.8
decorator==4.4.2
docutils==0.15.2
easydict==1.9
en-core-web-sm==3.0.0
filelock==3.0.12
gast==0.2.2
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==3.7.2
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.9.5
joblib==1.0.1
json-lines==0.5.0
jsonlines==1.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lmdb==0.94
Markdown==3.2.1
matplotlib==3.2.0
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.0
msgpack-numpy==0.4.4.3
networkx==2.4
numpy==1.19.2
olefile==0.46
opt-einsum==3.2.0
packaging==20.9
pandas==1.1.5
parso==0.8.0
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==8.0.1
pip==20.3.1
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
Pygments==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-transformers==1.1.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==19.0.2
regex==2020.2.20
requests==2.23.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.16.2
scipy==1.4.1
sentencepiece==0.1.85
setuptools==51.0.0.post20201207
six==1.15.0
stanza==1.2
tabulate==0.8.6
tensorboard==1.15.0
tensorboardX==1.8
tensorflow==1.15.2
tensorflow-estimator==1.15.1
tensorpack==0.9.4
termcolor==1.1.0
tokenizers==0.10.1
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.31.1
traitlets==4.3.3
transformers==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
Werkzeug==1.0.0
wheel==0.36.1
wrapt==1.12.1
zipp==3.4.1
